pyproject.toml
[tool.poetry]
name = "ts-api"
version = "0.1.0"
description = ""
authors = ["william villani stiehler <william.stiehler@olist.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
flake8 = "^3.9.2"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
